// packages/api/src/generated/user/models/getRecruitmentPostsSort.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */

export type GetRecruitmentPostsSort = typeof GetRecruitmentPostsSort[keyof typeof GetRecruitmentPostsSort];


export const GetRecruitmentPostsSort = {
  LATEST: 'LATEST',
  DEADLINE: 'DEADLINE',
  VIEW_COUNT: 'VIEW_COUNT',
  COMMENT_COUNT: 'COMMENT_COUNT',
} as const;