// packages/api/src/generated/user/models/recruitmentPostDetailResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */
import type { JsonNode } from './jsonNode';
import type { RecruitmentPostAuthorResponse } from './recruitmentPostAuthorResponse';
import type { RecruitmentPostContactResponse } from './recruitmentPostContactResponse';
import type { RecruitmentPostDetailResponsePositionsItem } from './recruitmentPostDetailResponsePositionsItem';
import type { RecruitmentPostDetailResponseProgressMethod } from './recruitmentPostDetailResponseProgressMethod';
import type { RecruitmentPostDetailResponseRecruitmentStatus } from './recruitmentPostDetailResponseRecruitmentStatus';
import type { RecruitmentPostDetailResponseRecruitmentType } from './recruitmentPostDetailResponseRecruitmentType';

export interface RecruitmentPostDetailResponse {
  id: number;
  author: RecruitmentPostAuthorResponse;
  title: string;
  /**
     * | 값 | code | 설명 |
     * | --- | --- | --- |
     * | `SIDE_PROJECT` | 1 | 사이드 프로젝트 |
     * | `STUDY` | 2 | 스터디 |
     */
  recruitmentType: RecruitmentPostDetailResponseRecruitmentType;
  /**
     * | 값 | code | 설명 |
     * | --- | --- | --- |
     * | `RECRUITING` | 1 | 모집 중 |
     * | `CLOSED` | 2 | 마감 |
     */
  recruitmentStatus: RecruitmentPostDetailResponseRecruitmentStatus;
  recruitmentStartDate: string;
  recruitmentEndDate: string;
  /**
     * | 값 | code | 설명 |
     * | --- | --- | --- |
     * | `ONLINE` | 1 | 온라인 |
     * | `OFFLINE` | 2 | 오프라인 |
     * | `HYBRID` | 3 | 온·오프라인 |
     */
  progressMethod: RecruitmentPostDetailResponseProgressMethod;
  capacity: number;
  activityDurationMonths: number;
  technologyStacks: string[];
  positions: RecruitmentPostDetailResponsePositionsItem[];
  contact: RecruitmentPostContactResponse;
  summary: string;
  content: JsonNode;
  eligibilityAndSelectionProcess?: string;
  viewCount: number;
  commentCount: number;
  bookmarkCount: number;
  bookmarked: boolean;
}