// packages/api/src/generated/user/models/successResponsePageResponseRecruitmentPostSummaryResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */
import type { PageResponseRecruitmentPostSummaryResponse } from './pageResponseRecruitmentPostSummaryResponse';

export interface SuccessResponsePageResponseRecruitmentPostSummaryResponse {
  status: number;
  message: string;
  data?: PageResponseRecruitmentPostSummaryResponse;
}