// packages/api/src/generated/user/models/successResponseRecruitmentPostDetailResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */
import type { RecruitmentPostDetailResponse } from './recruitmentPostDetailResponse';

export interface SuccessResponseRecruitmentPostDetailResponse {
  status: number;
  message: string;
  data?: RecruitmentPostDetailResponse;
}