// packages/api/src/generated/user/models/successResponsePageResponseUserBootcampSummaryResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */
import type { PageResponseUserBootcampSummaryResponse } from './pageResponseUserBootcampSummaryResponse';

export interface SuccessResponsePageResponseUserBootcampSummaryResponse {
  status: number;
  message: string;
  data?: PageResponseUserBootcampSummaryResponse;
}