// packages/api/src/generated/user/models/successResponsePageResponseUserJobSummaryResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */
import type { PageResponseUserJobSummaryResponse } from './pageResponseUserJobSummaryResponse';

export interface SuccessResponsePageResponseUserJobSummaryResponse {
  status: number;
  message: string;
  data?: PageResponseUserJobSummaryResponse;
}